add license and move it
