+++ core/src/main/java/org/opensearch/sql/calcite/udf/UserDefinedFunction.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package org.opensearch.sql.calcite.udf;
 
 public interface UserDefinedFunction {
