+++ core/src/main/java/org/opensearch/sql/calcite/udf/conditionUDF/IfFunction.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package org.opensearch.sql.calcite.udf.conditionUDF;
 
 import org.opensearch.sql.calcite.udf.UserDefinedFunction;
