+++ core/src/main/java/org/opensearch/sql/calcite/udf/conditionUDF/NullIfFunction.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package org.opensearch.sql.calcite.udf.conditionUDF;
 
 import java.util.Objects;
