+++ core/src/main/java/org/opensearch/sql/calcite/udf/mathUDF/CRC32Function.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package org.opensearch.sql.calcite.udf.mathUDF;
 
 import java.util.zip.CRC32;
